+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2026-01-08 03:20:38
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: 2b7f7fa42e22e983e6ed1c3a066b704060c18f39
+
+## [ForwardWidgets-2kuai] - 2026-01-08 03:20:38
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 2c8a7af8e22aca2872127c4278a8f60a0683b5e5
+
+## [ForwardWidgets-pack1r] - 2026-01-08 03:20:38
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 59e8b283fee8d745e6b57bf995ca76cdb6a943e0
+
+
 ## [ForwardWidgets-huangxd] - 2026-01-07 03:20:19
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main